+++ hooks/useBarcodeScanner.ts
@@ -12,17 +12,17 @@ import {
     areOriginalUrlAndHoppedSimilar,
     checkIfThereAreHops,
     checkIfIsValidURL,
-} from "@/services/checkUrl";
+} from "@/services/URLServices/checkUrl";
 import {
     checkIfDomainIsSuspicious,
     checkIfTLDIsRare
-} from "@/services/checkDomainAndSubdomain"
+} from "@/services/URLServices/checkDomainAndSubdomain"
 import {
     isUrlSafe,
     getUrlReportAnalysis,
     scanUrl,
     isUrlSuspicious
-} from "@/services/getUrlReport";
+} from "@/services/VirusTotalService/getUrlReport";
 
 export function useBarcodeScanner () {
     const throttleDelay = 2000;
