+++ hooks/useSearchParamsFromImage.ts
@@ -10,17 +10,17 @@ import {
     areOriginalUrlAndHoppedSimilar,
     checkIfThereAreHops,
     checkIfIsValidURL,
-} from "@/services/checkUrl";
+} from "@/services/URLServices/checkUrl";
 import {
     checkIfDomainIsSuspicious,
     checkIfTLDIsRare
-} from "@/services/checkDomainAndSubdomain";
+} from "@/services/URLServices/checkDomainAndSubdomain";
 import {
     isUrlSafe,
     getUrlReportAnalysis,
     scanUrl,
     isUrlSuspicious
-} from "@/services/getUrlReport";
+} from "@/services/VirusTotalService/getUrlReport";
 
 export function useSearchParamsFromImage () {
     const { uri, qrdata } = useLocalSearchParams();
